+++ Change_Log.md
@@ -50,4 +50,4 @@ Day 7 (12/6/2025):
 Day 8 (12/7/2025):
 - Added last intersection image
 - Added funtional jewelry story robbery with a chance to miss it and changing images.
--  
+- Added puting out the dumpster fire to get key and unlock front door